Fixed a user reported issue where directory that a symlink links to also get deleted during undeploy/redeploy. Added a check for symlink before we try to delete the directory. Ran QL, deployment dev tests.
git-svn-id: https://svn.java.net/svn/glassfish~svn/trunk/main@56089 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5

+++ nucleus/deployment/common/src/main/java/com/sun/enterprise/deploy/shared/FileArchive.java
@@ -541,21 +541,28 @@ public class FileArchive extends AbstractReadableArchive implements WritableArch
 
         boolean allDeletesSucceeded = true;
         // delete contents
-        File[] entries = directory.listFiles();
-        for (int i=0;i<entries.length;i++) {
-            if (entries[i].isDirectory()) {
-                allDeletesSucceeded &= deleteDir(entries[i]);
-            } else {
-                if ( ! entries[i].equals(StaleFileManager.Util.markerFile(archive))) {
-                    final boolean fileDeleteOK = FileUtils.deleteFile(entries[i]);
-                    if (fileDeleteOK) {
-                        staleFileManager.recordDeletedEntry(entries[i]);
+
+        /*
+        *Do not recursively delete the contents if the current directory
+        *is a symbolic link.
+        */
+        if (FileUtils.safeIsRealDirectory(directory)) {
+            File[] entries = directory.listFiles();
+            for (int i=0;i<entries.length;i++) {
+                if (entries[i].isDirectory()) {
+                    allDeletesSucceeded &= deleteDir(entries[i]);
+                } else {
+                    if ( ! entries[i].equals(StaleFileManager.Util.markerFile(archive))) {
+                        final boolean fileDeleteOK = FileUtils.deleteFile(entries[i]);
+                        if (fileDeleteOK) {
+                            staleFileManager.recordDeletedEntry(entries[i]);
+                        }
+                        allDeletesSucceeded &= fileDeleteOK;
                     }
-                    allDeletesSucceeded &= fileDeleteOK;
                 }
             }
         }
-        // delete self
+        // delete self (the directory or the symbolic link)
         return (allDeletesSucceeded && FileUtils.deleteFile(directory));
     }
